// lib/types.ts
import { ExecutionError, Result } from '@e2b/code-interpreter'

type ExecutionResultBase = {
  sbxId: string
}

export type ExecutionResultInterpreter = ExecutionResultBase & {
  template: string
  stdout: string[]
  stderr: string[]
  runtimeError?: ExecutionError
  cellResults: Result[]
}

export type ExecutionResultWeb = ExecutionResultBase & {
  template: string
  url: string
}

export type ExecutionResult = ExecutionResultInterpreter | ExecutionResultWeb

import { Message } from './messages'

export interface ChatSession {
  id: string;
  title: string;
  updatedAt: number;
  messages: Message[];
}